Fix the bad video issue on Window client by increasing the rtp recv buffer size. Send buffer doesn't really matter, just to keep the same as talk does.
The same fix is submitted to libjingle for reivew. But I think it's worth to fix it here too as
it may take while for webrtc to get from libjingle. This patch is slightly different then that
one as I don't want to add the webrtcvideoengine.h back to webrtc.
Review URL: http://webrtc-codereview.appspot.com/166002

git-svn-id: http://webrtc.googlecode.com/svn/trunk@634 4adac7df-926f-26a2-2b94-8c16560cd09d