Rolled back and re-adapted the codec test.
I don't really think this solves the problem and it makes the code worse, but it seems to make the vie_auto_test binary usable again. I've never seen it chrash on the standard tests since I rewrote it. It can still chrash on the API tests and extended tests though, which indicates that there is a fundamental design / threading problem somewhere, like we suspected. I think this is a kind of progress for now since we can use vie_auto_test to detect new errors again.

BUG=
TEST=

Review URL: https://webrtc-codereview.appspot.com/425003

git-svn-id: http://webrtc.googlecode.com/svn/trunk@1857 4adac7df-926f-26a2-2b94-8c16560cd09d