Fixed sync issues in network emulator script + fix lint complaints
Somehow a merge conflict was committed when I submitted
http://review.webrtc.org/1158006/ which resulted in
https://code.google.com/p/webrtc/source/detail?r=3639

Changed _run_ipfw_command from a method to a function to satisfy lint in network_emulator.py, removed "self." on the function calls.

Renamed msg to fail_msg in the _run_ipfw_command and __init__ to make it more clear that it contains a failure message.
Review URL: https://webrtc-codereview.appspot.com/1157005

git-svn-id: http://webrtc.googlecode.com/svn/trunk@3642 4adac7df-926f-26a2-2b94-8c16560cd09d