Document updating gflags and remove code duplication.
When rolling the google-gflags dependency, there might be
a need of updating the generated configuration files. I added
a instructions to the README.webrtc file for doing that.

This CL also removes duplicated configuration headers so we
only separete the ones that differs (Windows and everything
else).

BUG=2251
TEST=none
R=fischman@webrtc.org, niklas.enbom@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/2046004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@4574 4adac7df-926f-26a2-2b94-8c16560cd09d