GN: Fix include paths for WebRTC in Chromium build.
Most WebRTC source files are using full paths for includes which
requires the root to be in the include path.

This is currently handled in the common_inherited_config config in
webrtc/BUILD.gn: the .. include_dir.

However, when built from Chromium, the include
paths are not inherited in the same way when building the all target.
Building the 'webrtc' target of Chrome works without the changes
in this CL, but the default target fails.

BUG=3441
TEST=Built the default target from a Chromium checkout with
https://codereview.chromium.org/321313006/ applied and
src/third_party/webrtc linked to the webrtc folder of the WebRTC
workspace.

R=brettw@chromium.org

Review URL: https://webrtc-codereview.appspot.com/15989004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@6670 4adac7df-926f-26a2-2b94-8c16560cd09d