Remove unneeded WebKit dependency from DEPS.
This dependency was added in r4173 in order to support
the build/android/run_tests.py script.
In http://crrev.com/217588 that script was removed and
build/android/test_runner.py was the preferred script to
launch Android tests with.
Also in http://crrev.com/279515 the dependency on webkitpy
was removed and replaced by copies of the needed files in
build/android/pylib/utils/.

BUG=None
TEST=None
R=tommi@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/22139004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@6918 4adac7df-926f-26a2-2b94-8c16560cd09d