Fix clang -Wformat warnings.
An unsigned int was passed through %lu instead of %u (harmless on 32bit).
More seriously, a wide string was passed through %s, which means only the
first byte in the string got printed (since the 2nd byte is likely 0 in
UCS-2). Use %ls to include the whole string, even though it might not be
renderable in the target 8bit buffer.

BUG=chromium:82385
R=henrike@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/22409004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@6946 4adac7df-926f-26a2-2b94-8c16560cd09d