Disable VideoAdapterTest.BlackOutput on DrMemory.
Reports uninitialized-memory reads that seem to originate from when the
frame is copied. The test passes if we remove CPU optimizations from
libyuv, disabling test until we figure out whether it's an unsupported
instruction in DrMemory, bug in libyuv or bug in the test.

BUG=3754
R=aluebs@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/19149004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@6995 4adac7df-926f-26a2-2b94-8c16560cd09d