Enable the iSACfix AudioDecoder test (and make it work again)
As far as I can tell, the test should have been enabled again once
https://code.google.com/p/webrtc/issues/detail?id=1353 was fixed, but
it wasn't, and has rotted a bit as a result. I'm not sure why the
number of encoded bytes have changed, but the output seems to be
correct (EncodeDecodeTest encodes, decodes, and compares the result
with the original).

The DecodePlc change is necessary because r7912 added support for that
to the iSACfix AudioDecoder.

BUG=1353, 3926
R=henrik.lundin@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/28309004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@7927 4adac7df-926f-26a2-2b94-8c16560cd09d