Bye request should use POST not GET
AppRTCDemo is failing to cleanly exit a room because it sends a GET request to /bye. The request to /bye should be a POST request. Because the /bye request is failing, the room is still marked as "full" and rejoining will fail.

BUG=
R=tkchin@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/47759004

Patch from Chuck Hays <haysc@webrtc.org>.

Cr-Commit-Position: refs/heads/master@{#8860}