Revert "Add h.264 AVC SPS parsing for resolution."
The Mac64 Debug builder is broken for an unknown failure (trybot is
green, no failure obvious in the commit break). Reverting this CL to see
if it goes green again, and then relanding to see if it is just some
weird flaky build issue.

This reverts commit 5ea8eff55ec21a1d81aaf7d29c0106fe13256150.

BUG=
TBR=rollback

Review URL: https://webrtc-codereview.appspot.com/47019004

Cr-Commit-Position: refs/heads/master@{#9074}