Add a link to tools/vim to use the Chromium YCM config with webrtc.
The YCM plugin needs build configuration to enable the Clang-based
completion engine for C++ code. This is provided by the tools/vim
directory, but it expects to be in the same checkout as the file under
edit.

To use, follow the directions here:
https://code.google.com/p/chromium/codesearch#chromium/src/tools/vim/chromium.ycm_extra_conf.py

External YCM plugin:
https://github.com/Valloric/YouCompleteMe

google3 use:
go/ycm

R=kjellander@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/49359004

Cr-Commit-Position: refs/heads/master@{#9223}