rtpAnalyze matlab tool: filter out RTCP packets
This change relates to the matlab tool rtpAnalyze. With this change,
RTP packets with payload types 72 through 76 are removed. In IETF
RFC3551, section "Payload Type Definitions", this range is marked as
reserved so that RTCP and RTP packets can be reliably distinguished.

BUG=webrtc:2692
TBR=tina.legrand@webrtc.org
NOTRY=true

Review URL: https://codereview.webrtc.org/1284423006

Cr-Commit-Position: refs/heads/master@{#9724}