Fix VS 2015 warning by adding an additional cast
The OwningThread member of CRITICAL_SECTION is declared as having type
HANDLE but it is actually the thread's Thread ID which is a DWORD. When
doing 64-bit builds of Chromium with VS 2015 this triggers a warning
because of the suspicious conversion from 32-bit integer to 64-bit
pointer.

This change adds a cast (and some comments) to make the conversion
explicit and avoid the warning.

R=henrikg@webrtc.org
BUG=440500

Review URL: https://codereview.webrtc.org/1386183002

Cr-Commit-Position: refs/heads/master@{#10190}