Remove system_wrappers dep from field_trial_default and metrics_default.
field_trial_default and metrics_default don't use system_wrappers and
don't need to depend on it. The dependency on field_trial_default was
added in libjingle in crrev.com/356135 and that broke compilation of
libjingle for NaCl with GN because system_wrappers currently doesn't
compile for NaCl.

TBR=niklas.enbom@webrtc.org

Review URL: https://codereview.webrtc.org/1412003007

Cr-Commit-Position: refs/heads/master@{#10495}