VideoCapturerAndroid: Use one thread per startCapture()/stopCapture() session
Currently, VideoCapturerAndroid sets the thread and handler in the ctor
and clears them in dispose(). This CL sets the handler in startCapture()
instead and clears it in stopCapture(). The purpose is to prepare for
sending in the SurfaceTextureHelper in startCapture() instead of letting
VideoCapturerAndroid create it in the ctor.

All access to the handler is now synchronized by a lock, and all
Runnables are posted with a token so that they can be removed all at
once in stopCapture() to guarantee that no pending operation will be
executed after stopCapture().

BUG=webrtc:5519

Review URL: https://codereview.webrtc.org/1763673002

Cr-Commit-Position: refs/heads/master@{#11939}