VideoCapturerAndroid: Ignore erroneous startCaptureOnCameraThread calls instead of crashing
Fix a bug where startCaptureOnCameraThread() is called while the camera is already successfully running. It may happen in the scenario when startCapture() is called, but startCaptureOnCameraThread() fails and posts a retry, then stopCapture() is called and removeCallbacksAndMessages() fails to remove the pending retry, and then startCapture() is called successfully.

BUG=b/28181364
R=glaznev@webrtc.org

Review URL: https://codereview.webrtc.org/1967053002 .

Cr-Commit-Position: refs/heads/master@{#12697}