Fixing race condition in IncomingVideoStream.
This really only happens on the memcheck bot. But the issue is that
the render thread may be started before the timer is started on
the main thread, which incorrectly attaches the timer to the render
thread. Then a thread check assertion occurs when the timer is
stopped on the main thread.

Simply starting the timer before starting the render thread fixes this.

BUG=webrtc:6062
TBR=tommi@webrtc.org

Review URL: https://codereview.webrtc.org/2105013002 .

Cr-Commit-Position: refs/heads/master@{#13317}