Un-flaking TestSrtpError by using a fake clock.
This test verifies that SRTP errors are only signaled if a specific
amount of time has passed, and was using ProcessMessages(time) to
wait for an amount of time. But ProcessMessages may sometimes wait
for longer than the requested time (especially on TSAN bot, etc.).

BUG=webrtc:4690
R=tommi@webrtc.org

Review URL: https://codereview.webrtc.org/2184083004 .

Cr-Commit-Position: refs/heads/master@{#13597}