iOS: Update MB and JSON configs + enable Goma
Turns out that if mb_type is missing in the JSON, GYP is run the
traditional way instead of having the MB configuration decide.
This turns on MB for those builders.
See https://codereview.chromium.org/2194703002 for how Chromium
switched from GYP->GN.

The JSON environment for GYP and GN is only used during runhooks
step since there are scripts that key on some of these environment variables.
The actual build that is compiled is defined by the MB config, which
is now updated to have component=static_library everywhere for iOS.
With this CL, all configs gets a full GYP+GN environment.

When flipping bots over to GN, the following line will need to be added
in addition to changing mb_type:
"additional_compile_targets": [ "all" ],

Goma was also enabled for all builders to reduce compile time.

BUG=589510
NOTRY=True

Review-Url: https://codereview.webrtc.org/2239643002
Cr-Commit-Position: refs/heads/master@{#13775}