Remove GYP execution from DEPS hooks (gclient {sync,runhooks})
Due to the recent roll of chromium_revision in DEPS in
https://codereview.webrtc.org/2269953002/ an Android checkout
now experiences an error when running GYP (which is automatically
done as a hook when you run 'gclient {sync,runhooks}'.

This CL removes the GYP invocation entirely, as it would soon be
removed anyway as part of the GN migration.

All our bots are executing GYP/GN as part of the MB step
(called 'generate_build_files'), so they won't be affected.

BUG=webrtc:5949,webrtc:6258
NOTRY=True

Review-Url: https://codereview.webrtc.org/2280633002
Cr-Commit-Position: refs/heads/master@{#13920}