Increase timeout for flaky tests for ProcessThreadImpl
These tests were waiting for a thread to start in 100 ms. This seems to
be not enough on some Windows bots, as there were multiple failures
recently:
https://build.chromium.org/p/client.webrtc/builders/Win32%20Debug/builds/9188/steps/modules_unittests/logs/stdio
https://build.chromium.org/p/client.webrtc/builders/Win64%20Debug/builds/8794/steps/modules_unittests/logs/stdio

I have increased the timeout to 500 ms. This will not make successful
test runs slower, but will reduce the chance of a spurious failure.

R=perkj@webrtc.org

Review URL: https://codereview.webrtc.org/2308183002 .

Cr-Commit-Position: refs/heads/master@{#14096}